% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_changelog.R, R/add_ctu_header.R
\name{add_ctu_header}
\alias{add_ctu_header}
\title{Insert a change log section to the active document
This function is not intended to be used in any scripts.
It is intended to be used with qmd files.
Place the cursor where you want the change log and run the addin.}
\usage{
add_ctu_header()

add_ctu_header()
}
\value{
adds a change log section to the active document

adds CTU header to the active document
}
\description{
Insert a change log section to the active document
This function is not intended to be used in any scripts.
It is intended to be used with qmd files.
Place the cursor where you want the change log and run the addin.

Insert CTU standard header to the active document
This function is not intended to be used in any scripts.
}
