% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gtsummary_n.R
\name{add_gtsummary_n}
\alias{add_gtsummary_n}
\title{Insert Ns into gtsummary tables}
\usage{
add_gtsummary_n()
}
\value{
adds gtsummary code to the active document
}
\description{
gtsummary, by default has suboptimal presentation of number of observations,
especially in terms of missing data. This function adds the code at
the cursor location of the active document.
}
\details{
The first part should be placed after a gtsummary::tbl_summary call and the
add_by_n function definition should be moved to whereever R functions are
defined in your project

This code only needs to be used for grouped tables. For ungrouped tables,
use \code{add_overall() |> add_n()}.

Code modified from \href{https://stackoverflow.com/questions/68676585/gtsummary-split-add-n-between-groups}{a question on stackoverflow}
}
