% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloured_text.R
\name{coloured_text}
\alias{coloured_text}
\title{Colour specific text in a html document}
\usage{
coloured_text(colour, ...)
}
\arguments{
\item{colour}{colour to use for the content}

\item{...}{content to be coloured}
}
\description{
Colour specific text in a html document
}
\details{
In quarto/Rmarkdown documents, use the function in an inline R chunk
(\code{`r coloured_text(...)`}) or in an R chunk with the result set to \code{asis}
}
\examples{
coloured_text("red", "This text should be red")
# also works with multiple contents
coloured_text("red", "This text should be red", "this one too")
# For multiple colours, use multiple calls to the function:
coloured_text("red", "this is red,")
coloured_text("blue", "and this is blue!")
}
