% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.univ.packages.R
\name{install.univ.packages}
\alias{install.univ.packages}
\title{Install packages from an R-universe}
\usage{
install.univ.packages(x, univ = "https://ctu-bern.r-universe.dev")
}
\arguments{
\item{x}{package(s) to install}

\item{univ}{universe to install from}
}
\description{
R-universes are a little like mini-CRANs. Packages are generally sourced from
github and compiled into binaries for easy installation. To install them,
though, you have to specify the repository argument in the install.packages
function. This function removes that requirement.
}
\examples{

# install.univ.packages("randotools")

}
