% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStageSS.r
\name{OneStageSS}
\alias{OneStageSS}
\title{OneStageSS}
\usage{
OneStageSS(p0, p1, power = 0.8, n = NA, alpha = 0.05, nstart = 1)
}
\arguments{
\item{p0}{Proportion of events under H0}

\item{p1}{Proportion of events under H1}

\item{power}{Power, default=0.8}

\item{n}{Sample size, calculated if NA, otherwise power is calculated}

\item{alpha}{Alpha, default=0.05}

\item{nstart}{Number of patients to start iterations, default=1, can be increased to save time}
}
\value{
List with number of patients with events for a significant trial (r), number of patients (n), 
	alpha and power.
}
\description{
Calculates sample size or power for single-stage trials with a binary endpoint according to:
	Ahern, R. P., Sample size tables for exact single-stage phase II designs.
Statistics in medicine 20.6 (2001): 859-866.
}
\examples{
# Calculate sample size
OneStageSS(p0=0.5,p1=0.6,alpha=0.05,power=0.8)
# Calculate power
OneStageSS(p0=0.5,p1=0.6,alpha=0.05,n=100)

}
