% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_rando_code.R
\name{template_rando_code}
\alias{template_rando_code}
\title{Template randomisation list code}
\usage{
template_rando_code()
}
\value{
an R script
}
\description{
Open a file with template code to generate a randomisation list (and report).
}
\details{
This must be used within RStudio.
}
\examples{

# template_rando_code()
}
