% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_.R
\name{use_param_report_template}
\alias{use_param_report_template}
\title{Template files for a parameterized report}
\usage{
use_param_report_template(save_as, open = TRUE, ...)
}
\arguments{
\item{save_as}{filename to save the files as (do not include extension as multiple files are created)}

\item{open}{logical indicating whether to open the file}

\item{...}{other arguments passed to \code{use_template}}
}
\value{
saves two files, an Rmd and an R file
}
\description{
Parameterized reports allow to create reports in a loop. Examples might be
site specific quality reports or recruitment reports.
}
\details{
This function creates 2 files as a barebones template for such parameterized
reports. An Rmd file is created (based on the UNIBE template) and an R file
containing a small loop. Currently, the you have to supervise this loop, as
\code{use_ubreportclo} (which relies on \code{usethis::use_template}) will
not overwrite files without being given explicit permission.
}
\examples{
# use_param_report_template("param")
}
