% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_.R
\name{use_qmd_html}
\alias{use_qmd_html}
\alias{use_qmd_pres}
\alias{use_qmd_htmlrecruitment}
\alias{use_qmd_htmlsampsi}
\title{Use the CTU quarto html template
This function will download and optionally open the quarto markdown (qmd) file
and the other necessary files from github.}
\usage{
use_qmd_html(save_in = ".", open = TRUE)

use_qmd_pres(save_in = ".", open = TRUE)

use_qmd_htmlrecruitment(save_in = ".", open = TRUE)

use_qmd_htmlsampsi(save_in = ".", open = TRUE)
}
\arguments{
\item{save_in}{directory to save the main qmd in}

\item{open}{logical indicating whether to open the file}
}
\description{
Use the CTU quarto html template
This function will download and optionally open the quarto markdown (qmd) file
and the other necessary files from github.

Use the CTU quarto revealjs presentation template
This function will download and optionally open the quarto markdown (qmd) file
and the other necessary files from github.

Use the CTU quarto html recruitment report template
This function will download and optionally open the quarto markdown (qmd) file
and the other necessary files from github.

Use the CTU quarto html sample size report template
This function will download and optionally open the quarto markdown (qmd) file
and the other necessary files from github.
}
\section{Functions}{
\itemize{
\item \code{use_qmd_pres()}: revealjs Presentation Template

\item \code{use_qmd_htmlrecruitment()}: Recruitment Report Template

\item \code{use_qmd_htmlsampsi()}: Sample Size Report Template

}}
\examples{
# dir <- tempdir()
# file <- file.path(dir, "filename.qmd")
# use_qmd_html(file, FALSE)
# in practice at CTU:
# use_qmd_html("08_Reports_xx/ReportName.qmd", TRUE)
# dir <- tempdir()
# file <- file.path(dir, "filename.qmd")
# use_qmd_pres(file, FALSE)
# in practice at CTU:
# use_qmd_pres("08_Reports_xx/ReportName.qmd", TRUE)
# dir <- tempdir()
# file <- file.path(dir, "filename.qmd")
# use_qmd_html(file, FALSE)
# in practice at CTU:
# use_qmd_htmlrecruitment("08_Reports_xx/ReportName.qmd", TRUE)
# dir <- tempdir()
# file <- file.path(dir, "filename.qmd")
# use_qmd_html(file, FALSE)
# in practice at CTU:
# use_qmd_htmlrecruitment("08_Reports_xx/ReportName.qmd", TRUE)
}
