% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_.R
\name{use_qmd_pres}
\alias{use_qmd_pres}
\title{Use the CTU quarto revealjs presentation template
This function will download and optionally open the quarto markdown (qmd) file
and the other necessary files from github.}
\usage{
use_qmd_pres(save_as = "file.qmd", open = TRUE)
}
\arguments{
\item{save_as}{filename to save the main qmd as (other filenames are fixed to be compatible with the qmd)}

\item{open}{logical indicating whether to open the file}
}
\description{
Use the CTU quarto revealjs presentation template
This function will download and optionally open the quarto markdown (qmd) file
and the other necessary files from github.
}
\examples{
# dir <- tempdir()
# file <- file.path(dir, "filename.qmd")
# use_qmd_pres(file, FALSE)
# in practice at CTU:
# use_qmd_pres("08_Reports_xx/ReportName.qmd", TRUE)
}
