% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_.R
\name{use_quarto}
\alias{use_quarto}
\title{Get a quarto template from the CTU-Bern quarto repository}
\usage{
use_quarto(x, dir, open = TRUE)
}
\arguments{
\item{x}{branch name of the template in question}

\item{dir}{folder in which to save the template}

\item{open}{logical, whether to open the file}
}
\value{
a qmd file (and extension file)
}
\description{
Get a quarto template from the CTU-Bern quarto repository
}
\examples{
#use_quarto("html", ".")
}
