% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_.R
\name{use_recreport_template}
\alias{use_recreport_template}
\title{Open a new recruitment report}
\usage{
use_recreport_template(save_as, open = TRUE, ...)
}
\arguments{
\item{save_as}{filename to save the Rmd as}

\item{open}{logical indicating whether to open the file}

\item{...}{other arguments passed to \code{use_template}}
}
\value{
Creates an Rmd file in the \code{save_as} location
}
\description{
Open a new recruitment report
}
\examples{
# use_recreport_template("rr.Rmd")
}
