% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_.R
\name{use_ssreport_template}
\alias{use_ssreport_template}
\title{Open a new sample size report}
\usage{
use_ssreport_template(save_as, open = TRUE, ...)
}
\arguments{
\item{save_as}{filename to save the Rmd as}

\item{open}{logical indicating whether to open the file}

\item{...}{other arguments passed to \code{use_template}}
}
\value{

}
\description{
Open a new sample size report
}
\examples{
# use_ssreport_template("ss.Rmd")
}
