% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_.R
\name{use_ub_tex_template}
\alias{use_ub_tex_template}
\title{Open a new tex template allowing customization}
\usage{
use_ub_tex_template(save_as, open = TRUE, ...)
}
\arguments{
\item{save_as}{filename to save the tex file as (should be the same as in the Rmd header)}

\item{open}{logical indicating whether to open the file}

\item{...}{other arguments passed to \code{use_template}}
}
\value{

}
\description{
Open a new tex template allowing customization
}
\examples{
# use_recreport_template("rr.Rmd")
}
