% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_report_template.R
\name{use_ubreportclo}
\alias{use_ubreportclo}
\title{Title}
\usage{
use_ubreportclo(
  dir,
  projnum = "xxx",
  projname = "Project YYY",
  reporttype = "Type of report",
  version = "Version",
  sign = "Signature",
  email = "author.name@ctu.unibe.ch",
  job = "Senior Statistician",
  open = FALSE,
  ...
)
}
\arguments{
\item{dir}{Directory in which to create the clo file (should be the same as
the Rmd to be knit)}

\item{projnum}{Project number to appear in report}

\item{projname}{Project name to appear in report}

\item{...}{}
}
\value{

}
\description{
Title
}
