% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ctu_header.R
\name{add_ctu_header}
\alias{add_ctu_header}
\title{Insert CTU standard header to the active document
This function is not intended to be used in any scripts.}
\usage{
add_ctu_header()
}
\value{
adds CTU header to the active document
}
\description{
Insert CTU standard header to the active document
This function is not intended to be used in any scripts.
}
