% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_pdf.R
\name{report_pdf}
\alias{report_pdf}
\title{Generate PDF report}
\usage{
report_pdf(
  report_template = find_resource("report", "report.tex"),
  latex_engine = "pdflatex",
  ...
)
}
\arguments{
\item{report_template}{tex template to use}

\item{latex_engine}{tex engine to use (pdflatex, lualatex, xelatex)}

\item{...}{options passed to bookdown}
}
\description{
This is not intended to be run manually... rather via knitting a document
}
