% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unibeRed.R
\name{unibeRed}
\alias{unibeRed}
\title{UNIBE red
Get the official UNIBE red colour}
\usage{
unibeRed(alpha = 1)
}
\arguments{
\item{alpha}{alpha for transparency (0 to 1)}
}
\value{
colour code
}
\description{
UNIBE red
Get the official UNIBE red colour
}
\examples{
plot(1:5, col = unibeRed())
}
