% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_report_template.R
\name{use_ubreportclo}
\alias{use_ubreportclo}
\title{Generate the CLO file for use with the UNIBE tex template}
\usage{
use_ubreportclo(
  dir,
  projnum = "xxx",
  projname = "Project YYY",
  reporttype = "Type of report",
  version = "Version",
  sign = "Author name",
  email = "author.name@ctu.unibe.ch",
  job = "Senior Statistician",
  open = FALSE,
  ...
)
}
\arguments{
\item{dir}{Directory in which to create the clo file (should be the same as
the Rmd to be knit)}

\item{projnum}{Project number to appear in report}

\item{projname}{Project name to appear in report}

\item{reporttype}{The main title of the report}

\item{version}{Version number or similar}

\item{sign}{Your name as it should appear}

\item{email}{your email address}

\item{job}{your job title}

\item{open}{open the file or not}

\item{...}{other options passed to \code{use_template}}
}
\value{

}
\description{
The CLO file contains the information identifying the author, project and
institution (although we do not modify that). It is used by the UNIBE CLS
file to create the parts of the template.
}
\examples{
# use_ubreportclo(dir = "temp", # THE SAME DIR AS THE Rmd FILE
#                 # Personal info
#                 sign = "Alan",
#                 email = "alan.haynes@ctu.unibe.ch",
#                 job = "Senior Statistician",
#                 # Project info
#                 projnum = "1234",
#                 projname = "Project X",
#                 # Report info
#                 version = Sys.Date(),
#                 reporttype = "Recruitment report"
#                 )
}
